This does a few things for TVRec::FinishedRecording()  * It adds the job handling code and some other code that should be executed for all recordings to the method.  * It adds an integer parameter to track where FinishRecording is being called from and add more info to the VERBOSE macro.  * It adds tracking so that the portions that should only be run once (such as starting transcoding jobs) are only run once.
The last two were added because FinishedRecording() is run multiple times for some recordings.. As far as I can tell this only happens when there is a ring buffer change or when a recording is cancelled early (the normal case for the last recording in LiveTV). Anyway the debugging should help understand the cases where this happens better. And the tracking makes sure we don't do lots of unnecessary work.