Tries to avoid QThread::terminate() in pthread->QThread port of HouseKeeper threads [64d51365568348b91e30].
Killing threads this way can result in deadlocks and hence is heavily discouraged. We also limit when termination can occur to the MythSystem::Wait() call in the MFD thread.

We also wait on the threads to start after calling QThread::start(). This is required because the QThread::wait() wait() call will return immediately if the thread isn't running yet and we depend of QThread::wait() as a sort of poor man's pthread_join(). A QThread will blithely destruct itself while it's thread is still running, resulting in some "entertaining" backtraces.