Threading Scheduler cleanups. See [a5a2a9c0192f3ab76829b351ab42346006295361].
 * Gets rid of the use of QThread::terminate() in scheduler shutdown (unsafe).
 * Makes sure the thread is running when needed using a conditional.
 * Makes Scheduler a QThread itself since it was already a QObject.

The first change is for a problem that existed before the port, I just happened to catch it because I was doing a code review. The second change is for a problem introduced by the port (very unlikely to cause any problems in this case). The third change should not affect functionality at all.