This fixes the scheduler no-reschuler sleep in the [35fd63c] the Scheduler threading cleanups.
In the mythtv-rec branch we don't run an reschedules if the current reschedule is expected
to take longer than the time to the next recording start. This allows our recordings to start
on time more reliably than in trunk. But there was a bug in my last commit that would allow
us to get stuck in the no-reschedule state for much longer than we should be in it which causes
scheduled recording to either not happen or not happen on time.