Rework of [07dd9d], convert Linux Firewire recorder's pthread to QThread.
The most important change here is to release m_lock before waiting on the thread to exit, without this we will deadlock. Other changes were to properly wait for the thread to start, which just makes an unlikely race condition impossible. Finally, a usleep has been converted to a wait condition sleep so that shutdown can be initiated a little more quickly if we're sleeping when a shutdown is requested (a 50 ms speedup best case; 25 ms average).