Slight rework of [eaa146], converting FIFOWriter pthread's to QThread.
Blocks on thread startup + changes shutdown so that we don't block on the numpipe # thread shutdowns until we've told all of them to shut down first (speeding things up a little bit). Some locking is changed slightly to take advantage of QMutexLocker. Finally, we now don't wait for data in the fifo loop when we've already been told to shutdown.