Rework of [72f13d]. Converts NVR from pthreads to QThreads.
The biggest change from [72f13d] is that this fixes some breakage introduced by the Qt3->Qt4 port. The port introduced several extra locks where pauseLock should have been used.

Based on a grep of the code this appears to eliminate the last of the orphan locks introduced by the Qt4 port in the mythtv-rec2 branch. There are still 11 instances in the master branch.