Don't ping DB port for localhost.
Do not test to see if something is listening on the DB host and port if
the DB host is localhost.  This change is required on systems that have
localhost mapped to the ipv6 ::1 or that disable MySQL TCP/IP
connections, after the change at 859ad36a6 (and, for users that specify
127.0.0.1 for the DBHostName, the change at 3364c44e4).

Users who changed to the routable IP address, rather than using
localhost or 127.0.0.1 for the DBHostName while this was broken should
change back, now, to get best performance.

Fixes #9756.  Thanks to Chris Pinkham for finding the cause.