Tries to limit RemoteFile checking of mounted DVD/BD discs to only be done when absolutely necessary.
The problem is that RemoteFile checks can be very slow and as done in master checks for /VIDEO_TS and /BDMV in remote filesystems are sometimes done quite frequently. Since these are also performed in the Qt UI thread they also prevent the screen from being drawn. In mythtv-rec2 I just disabled these completely a long time ago, but since that isn't really an acceptible solution for everyone this changeset instead limits remotefile checks to when they might be necessary (just before we try to play back a DVD or BD or video file.) This could be further optimized by not checking if the file ends in known video file extension, and/or by adding a single backend command that checks if a path is a BD or DVD path. But as is this already avoids hundreds of calls backend queries in the UI thread in exhange for at most two just before playing back a video.