Convert the RecordingTypes to a templated class
There was far too much repeated code here.  Having this with a templated
class structure leads to easier to maintain code as there is only one place
where the code is rather than (so far) 4.  Additionally, it is based on the
QHash lookups rather than chains of if-elsif-else structures using string
compares (one of the least efficient things to do repeatedly!)