Fixes #10667. Bind to multicast address.
If multiple servers multicast content on the same port we can't receive it on multiple
recorders if bind to QHostAddress::Any. If we bind to the multicast address we don't
have this problem. Problem discovered and patched by Niels Ole Kirkeby.

Note: This also adds error handling for when a bind fails.